requirements.txt
# Web app
streamlit>=1.44,<2

# Environment and LLM access
python-dotenv>=1.0,<2
openai>=1.30,<2

# PDF preprocessing
PyPDF2>=3.0,<4
pymupdf>=1.24,<2

# Export
python-docx>=1.1,<2
